requirements.txt
# ======================================================================
# BekaaSense — Pinned dependencies
# Python 3.11 target. Pins are chosen for reproducibility.
# ======================================================================

# --- Web framework ---
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.6.0
whitenoise==6.8.2
gunicorn==23.0.0

# --- Data / ML ---
numpy==1.26.4
pandas==2.2.3
scipy==1.13.1
scikit-learn==1.5.2
xgboost==2.1.3
imbalanced-learn==0.12.4
shap==0.46.0
matplotlib==3.9.2
seaborn==0.13.2
statsmodels==0.14.4
joblib==1.4.2

# --- Excel / data ingestion ---
openpyxl==3.1.5
xlrd==2.0.1

# --- API utilities ---
python-dotenv==1.0.1
pydantic==2.9.2

# --- Testing ---
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0

# --- Dev / notebooks (optional, for local) ---
jupyter==1.1.1
ipykernel==6.29.5
